feat: change coordinator components packaging
