+++ coordinator/utils/mock_utils.go
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package impl
+package utils
 
 import "github.com/streamnative/oxia/coordinator/model"
